Update tx_search.js
Hello Noobish, I will soon publish a GPS plugin with which the web server and all plugins will be dynamically supplied with GPS data. In order for the location determination in the web server to work, the attached code adjustment is necessary. I only managed to find a reasonable solution via the websocket. Unfortunately, everything else about updating the variables didn't work. If you don't install the GPS plugin, nothing will change, as the coordinates are still read from the config at the beginning. Only when GPS data is received will it be overwritten. 73, Jens